Check file type in PDF2PJG and PDF2PNG
Fix compile errors in PDFSendToThunderbird and PDFViewFileAndDelete.

+++ plug-in/PDF2PNG/C#/PDF2PNG.cs
@@ -9,6 +9,7 @@ static class PDF2PNG
     public const  string WIN2PDF_COMPANY = "Dane Prairie Systems";
     public const  string WIN2PDF_PRODUCT = "Win2PDF";
     public const  string PNG_RESOLUTION = "PNGResolution";
+    public const string PLUG_IN_NAME = "Win2PDF PDF2PNG Plug-in";
 
     public const  int ERROR_LOCKED = 212;
     public const  int ERROR_INVALID_FUNCTION = 2;
@@ -26,7 +27,7 @@ static class PDF2PNG
 
             if (args.Length == 0 || resolution == "")
             {
-                resolution = Interaction.InputBox("Enter PNG Resolution:", WIN2PDF_PRODUCT, "100");
+                resolution = Interaction.InputBox("Enter PNG Resolution:", PLUG_IN_NAME, "100");
                 Interaction.SaveSetting(WIN2PDF_COMPANY, WIN2PDF_PRODUCT, PNG_RESOLUTION, resolution);
             }
             else if (args.Length == 1)
@@ -50,60 +51,66 @@ static class PDF2PNG
 
                 if (File.Exists(win2pdfcmdline))
                 {
-                    string pngname = System.IO.Path.Combine(Path.GetDirectoryName(args[0]),Path.GetFileNameWithoutExtension(args[0])) + ".png";
-                    // command line format is: pdffile pngfile pagenumber xresolution yresolution
-                    string arguments1 = string.Format("pdf2png \"{0}\" \"{1}\" 0 {2} {3}", args[0], pngname, resolution, resolution);
-
-                    ProcessStartInfo startInfo = new ProcessStartInfo(win2pdfcmdline);
+                    String ext = Path.GetExtension(args[0]).ToUpper();
+                    if (ext == ".PDF")
                     {
-                        var withBlock = startInfo;
-                        withBlock.Arguments = arguments1;
-                        withBlock.WindowStyle = ProcessWindowStyle.Hidden;
-                    }
+                        string pngname = System.IO.Path.Combine(Path.GetDirectoryName(args[0]), Path.GetFileNameWithoutExtension(args[0])) + ".png";
+                        // command line format is: pdffile pngfile pagenumber xresolution yresolution
+                        string arguments1 = string.Format("pdf2png \"{0}\" \"{1}\" 0 {2} {3}", args[0], pngname, resolution, resolution);
 
-                    // execute the print direct command
-                    newProc = System.Diagnostics.Process.Start(startInfo);
-                    newProc.WaitForExit();
-                    if (newProc.HasExited)
-                    {
-                        switch (newProc.ExitCode)
+                        ProcessStartInfo startInfo = new ProcessStartInfo(win2pdfcmdline);
                         {
-                            case ERROR_SUCCESS:
-                                {
-                                    //delete the PDF on success
-                                    File.Delete(args[0]);
-                                    break;
-                                }
+                            var withBlock = startInfo;
+                            withBlock.Arguments = arguments1;
+                            withBlock.WindowStyle = ProcessWindowStyle.Hidden;
+                        }
 
-                            case ERROR_FILE_NOT_FOUND:
-                                {
-                                    System.Windows.Forms.MessageBox.Show(string.Format("Win2PDF could not find file: {0}", win2pdfcmdline), WIN2PDF_PRODUCT, MessageBoxButtons.OK, MessageBoxIcon.Error);
-                                    break;
-                                }
+                        // execute the print direct command
+                        newProc = System.Diagnostics.Process.Start(startInfo);
+                        newProc.WaitForExit();
+                        if (newProc.HasExited)
+                        {
+                            switch (newProc.ExitCode)
+                            {
+                                case ERROR_SUCCESS:
+                                    {
+                                        //delete the PDF on success
+                                        File.Delete(args[0]);
+                                        break;
+                                    }
 
-                            case ERROR_ACCESS_DENIED:
-                                {
-                                    System.Windows.Forms.MessageBox.Show(string.Format("Win2PDF command line failed, make sure Win2PDF is licensed: {0}", win2pdfcmdline), WIN2PDF_PRODUCT, MessageBoxButtons.OK, MessageBoxIcon.Error);
-                                    break;
-                                }
+                                case ERROR_FILE_NOT_FOUND:
+                                    {
+                                        System.Windows.Forms.MessageBox.Show(string.Format("Win2PDF could not find file: {0}", win2pdfcmdline), PLUG_IN_NAME, MessageBoxButtons.OK, MessageBoxIcon.Error);
+                                        break;
+                                    }
 
-                            default:
-                                {
-                                    MessageBox.Show(string.Format("Win2PDF command line failed: {0}, error code {1}", win2pdfcmdline, newProc.ExitCode), WIN2PDF_PRODUCT, MessageBoxButtons.OK, MessageBoxIcon.Error);
-                                    break;
-                                }
+                                case ERROR_ACCESS_DENIED:
+                                    {
+                                        System.Windows.Forms.MessageBox.Show(string.Format("Win2PDF command line failed, make sure Win2PDF is licensed: {0}", win2pdfcmdline), PLUG_IN_NAME, MessageBoxButtons.OK, MessageBoxIcon.Error);
+                                        break;
+                                    }
+
+                                default:
+                                    {
+                                        MessageBox.Show(string.Format("Win2PDF command line failed: {0}, error code {1}", win2pdfcmdline, newProc.ExitCode), PLUG_IN_NAME, MessageBoxButtons.OK, MessageBoxIcon.Error);
+                                        break;
+                                    }
+                            }
                         }
                     }
+                    else
+                        System.Windows.Forms.MessageBox.Show(string.Format("Invalid save as type {0}. The save as type must be set to PDF", ext), PLUG_IN_NAME);
                 }
                 else
-                    System.Windows.Forms.MessageBox.Show(string.Format("Win2PDF Pro is not installed.  Download Win2PDF at https://www.win2pdf.com/download/"));
+                    System.Windows.Forms.MessageBox.Show(string.Format("Win2PDF is not installed.  Download Win2PDF at https://www.win2pdf.com/download/"));
             }
             else
                 MessageBox.Show("Invalid number of parameters");
         }
         catch (Exception ex)
         {
-            var exception_description = string.Format("Win2PDF plug-in exception {0}, stack {1}, targetsite {2}", ex.Message, ex.StackTrace, ex.TargetSite);
+            var exception_description = string.Format("{0} exception {1}, stack {2}, targetsite {3}", PLUG_IN_NAME, ex.Message, ex.StackTrace, ex.TargetSite);
             MessageBox.Show(exception_description);
             using (EventLog eventLog = new EventLog("Application"))
             {
